% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{argsEnum}
\alias{argsEnum}
\title{Define an enum for different modes; access with argsType$<enum_element>
ht https://stackoverflow.com/questions/33838392/enum-like-arguments-in-r}
\usage{
argsEnum()
}
\description{
TypeBool: TRUE/FALSE
TypeValue: any value expressed as "--arg=Value", "--arg Value", or "-a Value"
TypeMultiVal: TypeValue, but allowing multiple values to be stored (ie, keywords)
TypeCount: value increments each time the param is used. eg, -v -v yields 2
TypeRange: splits a TypeValue like "1:3" into two variables, "1" and "3"
TypePositional: required, final argument
}
