% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\docType{data}
\name{argsType}
\alias{argsType}
\title{enum for different param types; access with argsType$<enum_element>}
\format{
An object of class \code{list} of length 6.
}
\usage{
argsType
}
\description{
enum for different param types; access with argsType$<enum_element>
}
\keyword{datasets}
