% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{init_command_line_parser}
\alias{init_command_line_parser}
\title{Initialize command-line parsing}
\usage{
init_command_line_parser(script, desc, ver = NA)
}
\arguments{
\item{script}{name of the R script}

\item{desc}{description of the script}

\item{ver}{tool version number (string)}
}
\description{
Initialize command-line parsing
}
\examples{
init_command_line_parser('MyCheckbook.R','My checkbook program', '1.0.0')
}
