% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{parse_command_line}
\alias{parse_command_line}
\title{Parse the command line}
\usage{
parse_command_line(args)
}
\arguments{
\item{args}{command line arguments}
}
\value{
a list of parsed data
}
\description{
Parse the command line
}
\examples{
args <- commandArgs(trailingOnly = TRUE)
mydata <- parse_command_line(args)
# After parse_command_line(), access values as mydata$<var_name>, eg
# writeLines (paste("command:",mydata$command))
# writeLines (paste("subcommand:",mydata$subcmd))
# writeLines (paste("infile:", mydata$infile))
# writeLines (paste("outfile:",mydata$outfile))
}
