% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{parse_date}
\alias{parse_date}
\title{Parses a date in YYYY-MM-DD, YYYYMMDD, YYYY-MM or YYYY format}
\usage{
parse_date(d)
}
\arguments{
\item{d}{the date to parse (string)}
}
\value{
A tuple: c(y, m, d)
}
\description{
Parses a date in YYYY-MM-DD, YYYYMMDD, YYYY-MM or YYYY format
}
\examples{
parse_date("2019-12-31")
parse_date("2019-12")
parse_date("2019")
# parse_date("2019-13-31") # bad date!
}
