% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{reg_argument_list}
\alias{reg_argument_list}
\title{Register command line arguments}
\usage{
reg_argument_list(plist)
}
\arguments{
\item{plist}{list of arguments}
}
\description{
Register command line arguments
}
\examples{
arguments <- list(
  c("--outfile","-o","outfile",NA,argsType$TypeValue,'location of output file'),
  c("--date","-d","date",NA,argsType$TypeValue,'specify date'),
  c("--msg","-m","msg",NA,argsType$TypeValue,'memo line message'),
  c("--amount","-a","amount",NA,argsType$TypeValue,'specify dollar amount'),
  c("--payee","-p","payee",NA,argsType$TypeValue,'specify payee'))
reg_argument_list(arguments)
}
