% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{reg_positionals_list}
\alias{reg_positionals_list}
\title{Register a list of 'positional' arguments}
\usage{
reg_positionals_list(plist)
}
\arguments{
\item{plist}{list of positional arguments: variable name, help text}
}
\description{
Register a list of 'positional' arguments
}
\examples{
args <- list(c("infile","input file"))
}
