% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{reg_subcmd_list}
\alias{reg_subcmd_list}
\title{Register subcommands using a list}
\usage{
reg_subcmd_list(slist)
}
\arguments{
\item{slist}{list of subcommands}
}
\description{
Register subcommands using a list
}
\examples{
subcmds <- list(
c("paycheck", "deposit", "add a paycheck deposit"),
c("reimbursement", "deposit", "add a reimbursement"),
c("bankfee", "withdraw", "add a bank fee"),
c("check", "deposit", "add a check deposit")
)
reg_subcmd_list(subcmds)
}
