% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{usage}
\alias{usage}
\title{usage: Create a help message; adds --help and --ver params if not provided.}
\usage{
usage()
}
\description{
usage: Create a help message; adds --help and --ver params if not provided.
}
