% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{reg_command_list}
\alias{reg_command_list}
\title{Register commands using a list}
\usage{
reg_command_list(clist)
}
\arguments{
\item{clist}{list of lists of commands: command, help string}
}
\description{
Register commands using a list
}
\examples{
cmds <- list(
  list("withdraw", "add a withdrawal"),
  list("plot", "graph output"),
  list("deposit", "add a deposit"),
  list("edit", "update a record"),
  list("find", "find a record")
)
reg_command_list(cmds)
}
