% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command_line.R
\name{reg_subcmd_list}
\alias{reg_subcmd_list}
\title{Register subcommands using a list}
\usage{
reg_subcmd_list(slist)
}
\arguments{
\item{slist}{list of lists of subcommands: subcmd, parent, help string}
}
\description{
Register subcommands using a list
}
\examples{
subcmds <- list(
list("paycheck", "deposit", "add a paycheck deposit"),
list("reimbursement", "deposit", "add a reimbursement"),
list("bankfee", "withdraw", "add a bank fee"),
list("check", "deposit", "add a check deposit")
)
reg_subcmd_list(subcmds)
}
